% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_integrity.R
\name{check_data_integrity}
\alias{check_data_integrity}
\title{Check data integrity}
\usage{
check_data_integrity(x)
}
\arguments{
\item{x}{dataframe with at least colums lat, long and primary_fur_color}
}
\value{
Original dataframe if all tests are good. Otherwise stops.
}
\description{
Check data integrity
}
\examples{
datafile <- system.file("nyc_squirrels_sample.csv", package = "fusen.squirrels")
nyc_squirrels <- read.csv(datafile)
nyc_squirrels_ok <- check_data_integrity(nyc_squirrels)
}
